requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
prophet>=1.1.4
yfinance>=0.2.28
holidays>=0.34
tqdm>=4.66.1

# Technical Analysis
ta>=0.10.0  # Python Technical Analysis library

# Interactive CLI
questionary>=2.0.1

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Utilities
PyYAML>=6.0.1
python-dateutil>=2.8.2
pytz>=2023.3
joblib>=1.3.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.9.1
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.1
pylint>=3.0.1

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0
nbsphinx>=0.9.3
ipython>=8.16.1 